Got vim-devicons working on native linux
For some reason this plugin doesn't work on wsl... I will hopefully
figure it out soon. I have it working with a plugin that automatically
colors the glyphs for different file types. I also removed the buffer
movement mappings with space as a leader key. It breaks probaly because
of vim-fugitive, which is too good so I'll just stick with C-h C-j C-k
C-l for jumping across vim buffers and tmux instances